+++ packages/orq-rc/RELEASES.md
@@ -10679,3 +10679,13 @@ Based on:
 - [typescript v4.12.0-rc.12] packages/orq-rc
 ### Releases
 - [NPM v4.12.0-rc.12] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.12 - packages/orq-rc
+
+## 2026-07-10 07:08:27
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.789.1 (2.916.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.12.0-rc.13] packages/orq-rc
+### Releases
+- [NPM v4.12.0-rc.13] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.13 - packages/orq-rc